Migrate ProductPrice and ShopPayButton to Vitest
